AquaUtils: Add more device compatibility checks
AquariOS: Remove useless hasWifi check

From time to time we need to filter out certain features because not every
device is created equal. This makes it easy to call from one place instead
of adding the same code time and time again.

As of right now this class includes checks for

- Fingerprint support
- Fingerprint support & enrollment
- NFC
- Camera
- Bluetooth
- Mobile connection
- Alterative ambient display (OnePlus devices)
- A/B device support

Change-Id: I9e9555558455692356a09d8853abe31790bc70ae