AquaUtils: Add screen on action
Add an action to force the screen to turn on. This is a quick
and effective way to turn on the screen when we need it to. Intead
of hijacking WindowManagerPolicy in PWM, we instead call on PM to
do this.

- Make sure 'screen off' action doesn't work while screen is off.
This action should not be called if the screen is already off.

aosip edit: add volume panel action as well

Change-Id: Ie1774b2f318c9e8a21976d6ee43db0bbadff88bd