Add kill-app api (with new fresh implementation) [2/2]
We can kang what sysui assistant code does to check if launcher is running,
using a TaskStackChangeListener to always know the running
taskid and its component name. This is a solid and reliable way
to easily kill the task and even remove it from recents without the
risk to kill launcher or a different app than the foreground one.

Change-Id: I00c656f9fb68065ffe60b885284ab6ee084c66e4